backend/requirements.txt
# FastAPI and Server
fastapi==0.109.1
uvicorn==0.27.0
pydantic==2.6.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9

# Web Scraping
requests==2.31.0
beautifulsoup4==4.12.3
selenium==4.19.0
webdriver_manager==4.0.1
lxml

# ML and Data Processing
sentence-transformers==2.5.1
torch==2.2.0
pandas==2.2.0
numpy==1.26.3
scikit-learn==1.4.0

# Utilities
python-dotenv==1.0.0